implemented n-dimensional distance(point, line) function, all tests +engineering passed-->same behaviour

+++ src/Backend/painthelper.h
@@ -6,5 +6,5 @@ struct Pixel {
     unsigned char b, g, r;
 };
 
-Pixel getColorFromValue(float value, int maxValue=100);
-void paintpicture( std::vector<calc::line> lines, int width, int height, float zoomfactor=10, int maxDistColorGrad=10);
+Pixel getColorFromValue(double value, int maxValue=100);
+void paintpicture( std::vector<calc::line> lines, int width, int height, double zoomfactor=10, int maxDistColorGrad=10);
