+++ src/Backend/stringparser.cpp
@@ -18,9 +18,9 @@ std::vector<calc::SensorValue> getangles(const std::string& s) {
 
         for (const auto& item : j) {
             calc::SensorValue value;
-            value.theta = item["theta"].get<float>();
-            value.val = item["val"].get<float>();
-            value.xpos = item["xpos"].get<float>();
+            value.theta = item["theta"].get<double>();
+            value.val = item["val"].get<double>();
+            value.xpos = item["xpos"].get<double>();
             sensorValues.push_back(value);
         }
     } catch (const std::exception& e) {
