+++ src/Backend/stringparser.h
@@ -7,7 +7,7 @@
 #include <nlohmann/json.hpp>
 #include "calc.h"
 
-std::vector<float> getAnglesFromString(std::string str);
+std::vector<double> getAnglesFromString(std::string str);
 
 
 std::vector<calc::SensorValue> getangles(const std::string& s) ;
